refactor: move to default export everywhre for consistency
