--- app/ui/icons/bluesky.tsx
@@ -1,21 +1,0 @@
-interface IconProps {
-  size?: number | string;
-  className?: string;
-}
-
-export function Bluesky({ size = 24, className = "" }: IconProps) {
-  return (
-    <svg
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M5.76879 3.30388C8.29104 5.19742 11.004 9.03674 12.0001 11.0971C12.9962 9.03689 15.709 5.19738 18.2313 3.30388C20.0513 1.93757 23 0.8804 23 4.24437C23 4.9162 22.6148 9.8881 22.3889 10.6953C21.6036 13.5015 18.7421 14.2173 16.1967 13.7841C20.646 14.5413 21.7778 17.0496 19.3335 19.5579C14.6911 24.3216 12.661 18.3627 12.1406 16.8358C12.0453 16.5559 12.0007 16.4249 12 16.5363C11.9993 16.4249 11.9547 16.5559 11.8594 16.8358C11.3392 18.3627 9.30916 24.3218 4.66654 19.5579C2.22213 17.0496 3.35395 14.5412 7.80331 13.7841C5.25785 14.2173 2.39627 13.5015 1.6111 10.6953C1.38518 9.88802 1 4.91612 1 4.24437C1 0.8804 3.94882 1.93757 5.76866 3.30388H5.76879Z"
-        fill="#F6F5F4"
-      />
-    </svg>
-  );
-}
