--- app/util/nav-links.tsx
@@ -1,9 +1,0 @@
-export interface NavLinks {
-  path: string;
-  label: string;
-}
-
-export const paths = [
-  { path: "/", label: "Design" },
-  { path: "/photos", label: "Photos" },
-] as NavLinks[];
